requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
browsercookie==0.7.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.3.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.116
lz4==4.3.3
marshmallow==3.22.0
more-itertools==10.5.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.44.0
orjson==3.10.7
packaging==24.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
pyoverleaf==0.1.4
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
SecretStorage==3.3.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.11.0
zipp==3.20.1
